+++ RELEASE-NOTES.md
@@ -4,6 +4,17 @@ This is a pre-release version and APIs will change quickly. Before `1.0` release
 
 Please note after `1.0` Semver will be followed using normal protocols.
 
+# Version 0.10.0
+
+### Improvements
+* Each now supports else conditions
+* Each now supports custom index names {#each value, index in values}
+* Default index name is always `index` in data context instead of `@index`. `@foo` is not a valid variable name in js so it was breaking js expressions.
+
+### Bugs
+* You can now use signals like state in js expressions without get in templates i.e `{ someState == 'someValue' ? 'yes' : 'no' }` and not `{ someState.get() == 'someValue' ? 'yes' : 'no'}`
+* Fix `change` event to only fire on blur for ui-input, fix `input` event.
+
 # Version 0.9.4
 
 ### Bugs